refactor(core): cache pdp page data and metadata request
